+++ haystack_experimental/components/tools/openapi/openapi_tool.py
@@ -7,13 +7,13 @@ from pathlib import Path
 from typing import Any, Dict, List, Optional, Union
 
 from haystack import component, logging
+from haystack.components.generators.chat import OpenAIChatGenerator
 from haystack.dataclasses import ChatMessage, ChatRole
+from haystack.lazy_imports import LazyImport
 
-from haystack_experimental.components.tools.openapi.generator_factory import (
-    ChatGeneratorDescriptorManager,
-)
-from haystack_experimental.components.tools.openapi.openapi import (
+from haystack_experimental.components.tools.openapi._openapi import (
     ClientConfiguration,
+    LLMProvider,
     OpenAPIServiceClient,
 )
 
@@ -46,30 +46,28 @@ class OpenAPITool:
 
     def __init__(
         self,
-        model: str,
+        generator_api: LLMProvider,
+        generator_api_params: Dict[str, Any],
         tool_spec: Optional[Union[str, Path]] = None,
-        tool_credentials: Optional[Union[str, Dict[str, Any]]] = None,
-        model_kwargs: Optional[Dict[str, Any]] = None,
+        tool_credentials: Optional[str] = None,
     ):
         """
         Initialize the OpenAPITool component.
 
-        :param model: Name of the chat generator model to use.
+        :param generator_api: The API provider for the chat generator.
+        :param generator_api_params: Parameters for the chat generator.
         :param tool_spec: OpenAPI specification for the tool/service.
         :param tool_credentials: Credentials for the tool/service.
-        :param model_kwargs: Additional arguments for the chat generator model.
         """
-        manager = ChatGeneratorDescriptorManager()
-        self.descriptor, self.chat_generator = manager.create_generator(
-            model_name=model, **(model_kwargs or {})
-        )
+        self.generator_api = generator_api
+        self.chat_generator = self._init_generator(generator_api, generator_api_params)
         self.config_openapi: Optional[ClientConfiguration] = None
         self.open_api_service: Optional[OpenAPIServiceClient] = None
         if tool_spec:
             self.config_openapi = ClientConfiguration(
                 openapi_spec=tool_spec,
                 credentials=tool_credentials,
-                llm_provider=self.descriptor.name,
+                llm_provider=generator_api
             )
             self.open_api_service = OpenAPIServiceClient(self.config_openapi)
 
@@ -78,8 +76,8 @@ class OpenAPITool:
         self,
         messages: List[ChatMessage],
         fc_generator_kwargs: Optional[Dict[str, Any]] = None,
-        tool_spec: Optional[Union[str, Path, Dict[str, Any]]] = None,
-        tool_credentials: Optional[Union[dict, str]] = None,
+        tool_spec: Optional[Union[str, Path]] = None,
+        tool_credentials: Optional[str] = None,
     ) -> Dict[str, List[ChatMessage]]:
         """
         Invokes the underlying OpenAPI service/tool with the function calling payload generated by the chat generator.
@@ -104,7 +102,7 @@ class OpenAPITool:
             config_openapi = ClientConfiguration(
                 openapi_spec=tool_spec,
                 credentials=tool_credentials,
-                llm_provider=self.descriptor.name,
+                llm_provider=self.generator_api,
             )
             openapi_service = OpenAPIServiceClient(config_openapi)
 
@@ -134,4 +132,25 @@ class OpenAPITool:
             logger.error("Error invoking OpenAPI endpoint. Error: {e}", e=str(e))
             service_response = {"error": str(e)}
         response_messages = [ChatMessage.from_user(json.dumps(service_response))]
+
         return {"service_response": response_messages}
+
+    def _init_generator(self, generator_api: LLMProvider, generator_api_params: Dict[str, Any]):
+        """
+        Initialize the chat generator based on the specified API provider and parameters.
+        """
+        if generator_api == LLMProvider.OPENAI:
+            return OpenAIChatGenerator(**generator_api_params)
+        if generator_api == LLMProvider.ANTHROPIC:
+            with LazyImport("Run 'pip install anthropic-haystack'") as anthropic_import:
+                anthropic_import.check()
+                # pylint: disable=import-outside-toplevel
+                from haystack_integrations.components.generators.anthropic import AnthropicChatGenerator
+            return AnthropicChatGenerator(**generator_api_params)
+        if generator_api == LLMProvider.COHERE:
+            with LazyImport("Run 'pip install cohere-haystack'") as cohere_import:
+                cohere_import.check()
+                # pylint: disable=import-outside-toplevel
+                from haystack_integrations.components.generators.cohere import CohereChatGenerator
+            return CohereChatGenerator(**generator_api_params)
+        raise ValueError(f"Unsupported generator API: {generator_api}")
