+++ test/components/tools/openapi/test_openapi_client_edge_cases.py
@@ -5,7 +5,7 @@
 
 import pytest
 
-from haystack_experimental.components.tools.openapi.openapi import OpenAPIServiceClient, ClientConfiguration
+from haystack_experimental.components.tools.openapi._openapi import OpenAPIServiceClient, ClientConfiguration
 from test.components.tools.openapi.conftest import FastAPITestClient
